rename: AI -> AIManager in managers instead of utils
